use rosdoc_lite output_folder for cross referencing with doxygen
